# back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: "Ugram API"
  description: "A Ugram API documentation by team 12"
  version: "1.0"
host: "localhost:5000"
basePath: "/"
consumes:
- "application/json"
produces:
- "application/json"
...
schemes:
- "http"
paths:
...
  /person/id/{id_person}:
    get:
      tags:
      - "person"
      summary: "Get a user infos"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "The user details"
          schema:
            $ref: "#/definitions/ResponseObjectPerson"
  /person/all/{offset}:
    get:
      tags:
      - "person"
      summary: "Get all users infos"
      parameters:
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "An array of all users details"
          schema:
            $ref: "#/definitions/ResponseObjectPersonArray"
  /person/search/{search_text}/{offset}:
    get:
      tags:
      - "person"
      summary: "Search a user by forename, lastname or username"
      parameters:
      - in: "path"
        name: "search_text"
        required: true
        type: "string"
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "An array of user names"
          schema:
            $ref: "#/definitions/ResponseObjectPersonNameArray"
...
  /person/hashtag/{id_person}:
    get:
      tags:
      - "person"
      summary: "Recommand 5 users with common interests with the requested user"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
        description: "Id of the user asking for recommandations"
      responses:
        "200":
          description: "An array of users names"
          schema:
            $ref: "#/definitions/ResponseObjectPersonNameArray"
  /person/profile_picture/{id_person}:
    get:
      tags:
      - "person"
      summary: "Get a user profile picture"
      produces:
      - "image/png"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Image file"
          schema:
            type: "file"
    post:
      tags:
      - "person"
      summary: "Upload a profile picture"
      consumes:
       - "multipart/form-data"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      - in: "formData"
        name: "image"
        type: "file"
      responses:
        "200":
          description: "OK"
  /person:
    post:
      tags:
      - "person"
      summary: "Create a new user/account"
      parameters:
      - in: "body"
        name: "body"
        description: "User infos to create an account"
        required: true
        schema:
          $ref: "#/definitions/PersonInputCreate"
      responses:
        "201":
          description: "Id of the newly created user"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
    put:
      tags:
      - "person"
      summary: "Update a user/account infos"
      parameters:
      - in: "body"
        name: "body"
        description: "User infos to update"
        required: true
        schema:
          $ref: "#/definitions/PersonInputUpdate"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
    delete:
      tags:
      - "person"
      summary: "Delete a user/account"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the user to delete"
        required: true
        schema:
          $ref: "#/definitions/PersonInputDelete"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
...
  /post/person/{id_person}/{offset}:
    get:
      tags:
      - "post"
      summary: "Get all posts related to a user"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      - in: "path"
        name: "offset"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Array of posts"
          schema:
            $ref: "#/definitions/ResponseObjectPostArray"
...
  /notification/count/{id_person}:
    get:
      tags:
      - "notification"
      summary: "Get a person notifications number"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Notifications number"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
  /notification/id/{id_person}:
    get:
      tags:
      - "notification"
      summary: "Get a person notifications"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Notifications"
          schema:
            $ref: "#/definitions/ResponseObjectNotificationArray"
  /notification:
    put:
      tags:
      - "notification"
      summary: "Read a notification"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the notification to read"
        required: true
        schema:
          $ref: "#/definitions/NotificationInputUpdate"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
definitions:
...
  PersonInputCreate:
    type: "object"
    properties:
      lastname_person:
        type: "string"
        default: "Doe"
      firstname_person:
        type: "string"
        default: "John"
      gender_person:
        type: "string"
        enum:
        - "M"
        - "F"
        default: "M"
      birthdate_person:
        type: "string"
        format: "date"
        default: "2000-01-01"
      username_person:
        type: "string"
        default: "thejohndoe_"
      mail_person:
        type: "string"
        default: "john.doe@gmail.com"
      password_person:
        type: "string"
        format: "password"
        default: "password"
      phone_person:
        type: "string"
        default: "0152486359"
  PersonInputUpdate:
    type: "object"
    properties:
      lastname_person:
        type: "string"
        default: "Doe"
      firstname_person:
        type: "string"
        default: "John"
      mail_person:
        type: "string"
        default: "john.doe@gmail.com"
      phone_person:
        type: "string"
        default: "0152486359"
  PersonInputDelete:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
  PersonOutput:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      lastname_person:
        type: "string"
        default: "Doe"
      firstname_person:
        type: "string"
        default: "John"
      gender_person:
        type: "string"
        enum:
        - "M"
        - "F"
        default: "M"
      birthdate_person:
        type: "string"
        format: "date"
        default: "2000-01-01"
      username_person:
        type: "string"
        default: "thejohndoe_"
      mail_person:
        type: "string"
        default: "john.doe@gmail.com"
      phone_person:
        type: "string"
        default: "0152486359"
      confidentiality_person:
        type: "string"
        enum:
        - "PUBLIC"
        - "PRIVATE"
        default: "PUBLIC"
      displayonline_person:
        type: "boolean"
        default: "true"
      followersnumber_person:
        type: "integer"
        default: "0"
      followingnumber_person:
        type: "integer"
        default: "0"
      bio_person:
        type: "string"
        default: "..."
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
  PersonName:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      lastname_person:
        type: "string"
        default: "Doe"
      firstname_person:
        type: "string"
        default: "John"
      username_person:
        type: "string"
        default: "thejohndoe_"
...
  PostOutput:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      person:
        $ref: "#/definitions/PersonName"
      isliked_post:
        type: "boolean"
        default: "true"
      likes_post:
        type: "integer"
        default: "3"
      comments_post:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "My trip in Italy"
      hashtags_post:
        type: "string"
        default: "#italy"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
...
  NotificationInputUpdate:
    type: "object"
    properties:
      id_notification:
        type: "integer"
        default: "1"
  NotificationOutput:
    type: "object"
    properties:
      id_notification:
        type: "integer"
        default: "1"
      id_person_receiving:
        type: "integer"
        default: "1"
      person_sending:
        $ref: "#/definitions/PersonName"
      type_notification:
        type: "string"
        enum:
        - "LIKEPOST"
        - "COMMENT"
        - "LIKECOMMENT"
        - "FOLLOWER"
        default: "LIKEPOST"
      isread_notification:
        type: "boolean"
        default: false
      post:
        $ref: "#/definitions/NotificationPost"
      id_comment:
        type: "integer"
        default: "1"
      created_at:
        type: "string"
        format: "date-time"
  NotificationPost:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "..."
  ResponseObjectEmpty:
    type: "object"
    properties:
      response:
        type: "object"
        default: null
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectInt:
    type: "object"
    properties:
      response:
        type: "integer"
        default: 1
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
  ResponseObjectPerson:
    type: "object"
    properties:
      response:
        $ref: "#/definitions/PersonOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
  ResponseObjectPersonArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/PersonOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectPersonNameArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/PersonName"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectPostArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/PostOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"  
...
  ResponseObjectNotificationArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/NotificationOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"